Add cast type error

+++ src/Message/Command/Admin/Priority/Handler/GeneratePrioritiesHandler.php
@@ -34,7 +34,7 @@ readonly class GeneratePrioritiesHandler
             $priority = new Priority(
                 name: $jiraPriority->name,
                 description: $jiraPriority->description,
-                jiraId: $jiraPriority->id,
+                jiraId: (int) ($jiraPriority->id),
                 iconUrl: $jiraPriority->iconUrl,
                 statusColor: $jiraPriority->statusColor,
             );
